refactor: split index into modules
